views/user_views.py
from django.shortcuts import render
from rest_framework import status
from rest_framework.decorators import api_view
from rest_framework.response import Response

from forms import TranscriptionForm  # Make sure to adjust the import paths


def index(request):
    form = TranscriptionForm(request.POST or None, request.FILES or None)
    if request.method == 'POST':
        if form.is_valid():
            form.save()
            # Perform Whisper JAX API request for transcription
            # Store transcription request in the database

    return render(request, 'yourappname/index.html', {'form': form})

# ... other view functions ...


@api_view(['POST'])
async def async_transcription(request):
    # Process transcription asynchronously
    # Example: use asyncio or an async-friendly library for the transcription

    return Response({'message': 'Transcription request submitted'}, status=status.HTTP_202_ACCEPTED)
