requirements.txt
asgiref==3.7.2
click==8.1.7
colorama==0.4.6
Django==4.2.4
djangorestframework==3.14.0
joblib==1.3.2
nltk==3.8.1
psycopg2==2.9.7
pytz==2023.3
regex==2023.8.8
sqlparse==0.4.4
tqdm==4.66.1
tzdata==2023.3
whitenoise==6.5.0
